Finally get out of the stuck.
The error was in the branch instruction and the PC.
The branch just set the PC to the new address, but I was taking that
number as an offset in the rom, multiplied by 4 (because the addressing
is in 4 bytes). That is, instead of jumping from 0x0010 to 0x0020,
the jump was made from  0x0010 + (0x10 * 4) -> due to (u32 *)
addressing.

Add misc/ folder that contains dthumb disassembler to help
me verify if the decoding I was doing was right.

Add gba in the root folder that was using for testing.