Process INSTRUCTION_SOFTWARE_INTERRUPT (just changing states).
The SWI instruction involves executing code in the bios, but I don't
have that yet, resulting in every instruction being 0, so in goes in a
loop and crashes when it reaches invalid memory location.

Also change the get_instruction_at() methods (for ARM and THUMB)
because with SWI it can't loop up the instruction in the cartridge.