Change the way INSTRUCTION_LOAD_STORE_WITH_IMMEDIATE_OFFSET get address.
At 0x000009F0, the simulation fails because the address it was trying to
get the data from is invalid (r1 = 0x4000400 plus the offset 0x10, which
is 0x4000410). That address is in unused portion of memory by the GBA.
Apparently it's OK trying to access this memory, just do not do
anything. So the partial (maybe permanent) solution is to copy the
`get_memory_at` function to remove the asserts and just return 0; in the
calling function just verify that the address is valid before using it.

With this change the program runs until it fails for an unimplemented
instruction, so I call it a win! :D