Just to mark a checkpoint: time to implement cycle count.
I was making a huge mistake regarding the drawing; it has to be done
every frame (not every instruction like I do right now), that's why it's
toooo slow. The refresh rate is 280896, so I have to count how many
cycles each instruction has to add up to a counter and then begin
drawing.