Fix some cycles count and the display rate!!!
I was emulating every pipeline every frame of the window (setted at 60
FPS). That's why it was too slow. Trying to set the target FPS to 144,
200, or 300 improve the overall speed (until it caps at a certain speed,
sadly is not free).

Trying to draw every frame every time the GBA CPU reaches the frame
count works for a while, but there was a problem too: I could not paused
the frames because it was not rendered every frame (I think that was the
problem).

The key thing here is that I modified the run() method to emulate until
the GBA CPU reaches a frame, only then render the screen frame. This
saves a ton of time and I think it is the correct behaviour.