Finally make some progress on infinite loop 0x1774.
The loop is while waiting for the scanlines to be 159 (The max value).
To make it work, it is necessary update the io register 0x04000006 to
the current scanline. That's the value the loop at 0x1774 is looking for
(it breaks when the value is equal to 159).

Only saw it while dissasembly the bios in ghidra.