Showing the demo game!!!!
Finally I could make it happen!
The demo game is easy to display because is in mode 3, which means the
frames are stored in VRAM as bitmaps (just the exact pixel values),
therefore I don't have to look for palettes, or sprites or whatever
other GBA video thing is required.

But this is excellent progress :D